+++ CHANGELOG.md
@@ -4,3 +4,13 @@
 
 ### Added
 - Version Inicial
+
+## dev
+
+### Added
+ - Se agrega la seccion "rest" con controladores, excepciones, filtros y serializers
+ - Se agrega la seccion "db" con modelos, excepciones y configuracion para doctrine
+
+### Changed
+ - Se mejora la libreria ApiRestParams moviendola a lib/Filters/StandardFilter y deprecando la anterior
+ - Se mejora la libreria Pager moviendola a lib/Pagination/StandardPagination y deprecando la anterior